Moved logic out of errors replacer implementation class and into replacer super class
